[Reeder] hide unread badge & default browser
Since countless information or articles out there on the Internet, consuming them all is impossible, the unread badge is not a meaningful indicator for me.

On the other hand, the links opening from the external browser can keep us the browsing history, and Read Later articles also can be kept track of in one place in the Browser.